what's also annoying is this "do you really want to leave" confirmation when I close firefox. I suggest, that you only show it iff there is more than one unsaved worksheet. If all of them are saved, it's not an issue, right?-- Harald Schilly;Eenable it if some sync is failing for a file.

+++ salvus/page/exit_confirmation.coffee
@@ -4,18 +4,21 @@
 # code works in Firefox, Chrome, Safari and maybe IE, but does not seem to
 # work in Opera.  So Opera users could loose work...
 #
-# Also, this does not work on iOS.
+# Also, this does not work on i OS.
 #
 ###
 
 {top_navbar} = require('top_navbar')
+{unsynced_docs} = require('syncdoc')
 
 window.onbeforeunload = (e) ->
-    #
+    if not unsynced_docs()
+        return
+
     # TODO: we could instead just save everything at this point and return null
     #
     e.cancelBubble = true  # e.cancelBubble is supported by IE - this will kill the bubbling process.
-    mesg = ''#Leave the SageMath cloud?'
+    mesg = "Some documents haven't successfully synchronized with the server yet.  Leaving now may result in lost work."
     e.returnValue = mesg
     if e.stopPropagation
         e.stopPropagation()
